Display payments based on students list
This modifies the payments list to display a list of students first
then iterate over all the payments made by each student for each month.
I tried to avoid N+1 as much as possible although if given more time I
would come up with a better approach to this problem.